+++ Husky/Manager/StoreProvider.swift
@@ -39,7 +39,7 @@ class StoreProvider {
                 guard let name = info[Store.Schema.name] as? String else { return }
                 guard let latitude = info[Store.Schema.latitude] as? CLLocationDegrees else { return }
                 guard let longitude = info[Store.Schema.longitude] as? CLLocationDegrees else { return }
-                guard let scoredPeople = info[Store.Schema.scoredPeople] as? Double else { return }
+                guard let scoredPeople = info[Store.Schema.scoredPeople] as? Int else { return }
                 guard let storeScoreAverage = info[Store.Schema.storeScoreAverage] as? Double else { return }
                 stores.append(
                     Store(
